Generate AUTOSAR_00042.rnc with a more complete version
This was generated from AUTOSAR_00042_COMPACT.xsd in a 2 step process - first
was converted to rng via XMLtoRNG.xsl, and then that was converted to rnc using
trang.

The resulting rnc was then hand fixed-up to work as well as I can test it - is
valid according to jing, emacs nxml-mode is happy with it too and it says all
the few existing arxml files I have are valid as well.